+++ test/actions/BackendAction-test.js
@@ -5,7 +5,6 @@ const fsEx = require('fs-extra')
 const proxyquire = require('proxyquire')
 const mockFs = require('mock-fs')
 
-const utils = require('../../lib/utils/utils')
 const { SETTINGS_FOLDER, EXTENSIONS_FOLDER } = require('../../lib/app/Constants')
 const AppSettings = require('../../lib/app/AppSettings')
 const UserSettings = require('../../lib/user/UserSettings')
@@ -18,14 +17,31 @@ describe('BackendAction', () => {
    * @type {BackendAction}
    */
   let subjectUnderTest
-  let logger = {}
+
+  const utils = proxyquire('../../lib/utils/utils', {
+    '../logger': {
+      info: (param) => { info(param) },
+      error: (param) => { error(param) },
+      debug: (param) => { debug(param) }
+    }
+  })
+
   const BackendAction = proxyquire('../../lib/actions/BackendAction', {
-    '../logger': logger
+    '../logger': {
+      info: (param) => { info(param) },
+      error: (param) => { error(param) },
+      debug: (param) => { debug(param) }
+    },
+    '../utils/utils': utils
   })
 
   let appSettings
   let userSettings
   let dcHttpClient
+
+  let info
+  let error
+  let debug
 
   before(async () => {
     process.env.USER_PATH = userSettingsFolder
@@ -54,6 +70,7 @@ describe('BackendAction', () => {
       close: () => {},
       on: (event, fn) => fn(event, path.join(subjectUnderTest.pipelinesFolder, 'testPipeline.json'))
     }
+
     subjectUnderTest.extensionConfigWatcher = {
       start: sinon.stub().resolves(),
       on: sinon.stub().resolves(),
@@ -71,14 +88,15 @@ describe('BackendAction', () => {
 
     subjectUnderTest.dcHttpClient = {}
 
-    logger.info = () => {}
-    logger.error = () => {}
-    logger.debug = () => {}
+    info = (param) => {}
+    error = (param) => {}
+    debug = (param) => {}
   })
 
   afterEach(async () => {
     await subjectUnderTest.pipelineWatcher.close()
     await subjectUnderTest.extensionConfigWatcher.stop()
+    process.env.IGNORE_EXT_CONFIG_FOR = ''
   })
 
   after(async () => {
@@ -131,7 +149,7 @@ describe('BackendAction', () => {
       subjectUnderTest._pipelineRemoved = sinon.stub().resolves()
       subjectUnderTest._pipelineChanged = sinon.stub().resolves()
 
-      logger.debug = (msg) => {
+      debug = (msg) => {
         assert.equal(msg, 'The extension of the pipeline is not attached --> skip')
         assert.ok(!subjectUnderTest._pipelineRemoved.called, '_pipelineRemoved should not be called')
         assert.ok(!subjectUnderTest._pipelineChanged.called, '_pipelineChanged should not be called')
@@ -142,10 +160,16 @@ describe('BackendAction', () => {
   })
 
   describe('watching', () => {
-    const originalGetBlacklistedExtensions = utils.getBlacklistedExtensions
+    beforeEach(() => {
+      subjectUnderTest.dcHttpClient = {
+        downloadPipelines: sinon.stub().resolves({ pipelines: [{ pipeline: { id: 'testPipeline' } }] }),
+        removePipeline: sinon.stub().resolves(),
+        clearHooks: sinon.stub().resolves(),
+        uploadMultiplePipelines: sinon.stub().resolves()
+      }
 
-    after(() => {
-      utils.getBlacklistedExtensions = originalGetBlacklistedExtensions
+      subjectUnderTest.writeExtensionConfigs = () => { }
+      subjectUnderTest.appSettings.loadAttachedExtensions = () => { return { } }
     })
 
     it('should update pipelines', (done) => {
@@ -156,12 +180,6 @@ describe('BackendAction', () => {
         resetHooks: sinon.stub().resolves(),
         startStepExecutor: sinon.stub().resolves(),
         reloadPipelineController: sinon.stub().resolves()
-      }
-
-      subjectUnderTest.dcHttpClient = {
-        downloadPipelines: sinon.stub().resolves([{ pipeline: { id: 'testPipeline' } }]),
-        removePipeline: sinon.stub().resolves(),
-        uploadMultiplePipelines: sinon.stub().resolves()
       }
 
       subjectUnderTest.attachedExtensionsWatcher = {
@@ -182,7 +200,6 @@ describe('BackendAction', () => {
 
     it('should fail when pipeline IDs not matching pipeline file names', (done) => {
       appSettings.loadAttachedExtensions = () => { return { testExtension: { path: '..' } } }
-
       subjectUnderTest.backendProcess = {
         connect: sinon.stub().resolves(),
         selectApplication: sinon.stub().resolves(),
@@ -193,7 +210,7 @@ describe('BackendAction', () => {
       }
 
       subjectUnderTest.dcHttpClient = {
-        downloadPipelines: sinon.stub().resolves([]),
+        downloadPipelines: sinon.stub().resolves({ pipelines: [] }),
         removePipeline: sinon.stub().resolves(),
         uploadMultiplePipelines: sinon.stub().resolves()
       }
@@ -242,7 +259,7 @@ describe('BackendAction', () => {
       const file = path.join(subjectUnderTest.pipelinesFolder, 'plFooBarline1.json')
       assert.equal(subjectUnderTest.pipelines[file], undefined)
 
-      subjectUnderTest.dcHttpClient.uploadPipeline = (f, aId) => {
+      subjectUnderTest.dcHttpClient.uploadPipeline = async (f, aId) => {
         assert.deepEqual(subjectUnderTest.pipelines[file].id, pipeline.pipeline.id)
         assert.deepEqual(f, pipeline)
         assert.equal(aId, appId)
@@ -277,20 +294,19 @@ describe('BackendAction', () => {
 
     it('should not write generated extension-config if backend-extension was updated but extension is on blacklist', async () => {
       let generated = { id: 'myIgnoredExtension', backend: { id: 'myIgnoredExtension' } }
+      process.env.IGNORE_EXT_CONFIG_FOR = 'myIgnoredExtension'
 
-      let called = 0
-      subjectUnderTest.dcHttpClient.generateExtensionConfig = () => Promise.resolve()
-      utils.getBlacklistedExtensions = () => { return ['myIgnoredExtension'] }
-
-      logger.info = (msg) => {
+      let calledInfo = false
+      info = (msg) => {
         assert.equal('Ignoring extension-config.json of myIgnoredExtension (blacklisted)', msg)
+        calledInfo = true
       }
 
       const cfgPath = path.join(appPath, 'extensions', 'testExt')
 
       try {
         await utils.updateExtensionConfig({ file: generated, path: cfgPath }, await subjectUnderTest.appSettings.getId(), subjectUnderTest.dcHttpClient)
-        assert.equal(called, 0)
+        assert.ok(calledInfo)
       } catch (err) {
         assert.ifError(err)
       }
@@ -340,7 +356,8 @@ describe('BackendAction', () => {
       const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest3.json')
 
       subjectUnderTest.dcHttpClient = {
-        uploadPipeline: sinon.stub().resolves()
+        uploadPipeline: sinon.stub().resolves(),
+        downloadPipelines: sinon.stub().resolves({ pipelines: [{ pipeline: { id: 'testPipeline' } }] })
       }
 
       subjectUnderTest.backendProcess = {
@@ -405,18 +422,17 @@ describe('BackendAction', () => {
     it('should return if pipeline was removed', (done) => {
       const pipelineId = 'dCPlTest4'
       let called = false
-      subjectUnderTest._writeLocalPipelines = sinon.mock().resolves()
       subjectUnderTest.dcHttpClient.removePipeline = (plId, id, trusted) => {
         assert.equal(plId, pipelineId)
         called = true
       }
+      subjectUnderTest.dcHttpClient.downloadPipelines = sinon.stub().resolves({ pipelines: [{ pipeline: { id: 'testPipeline' } }] })
 
       const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest4.json')
       subjectUnderTest.pipelines[file] = { id: pipelineId }
 
       subjectUnderTest._pipelineRemoved(file).then(() => {
         assert.ok(called)
-        assert.ok(subjectUnderTest._writeLocalPipelines.called)
         done()
       }).catch(err => {
         assert.ifError(err)
@@ -424,6 +440,7 @@ describe('BackendAction', () => {
     })
 
     it('should work', async () => {
+      subjectUnderTest.pushHooks = () => {}
       appSettings.loadAttachedExtensions = () => { return { testExtension: { path: '..' } } }
       subjectUnderTest._startSubProcess = () => {}
       try {
@@ -447,9 +464,9 @@ describe('BackendAction', () => {
       await fsEx.writeJson(extensionConfigPath, mockConf)
       let called = 0
       subjectUnderTest.appSettings.getId = () => 1
-      subjectUnderTest.dcHttpClient.generateExtensionConfig = (config) => {
+      subjectUnderTest.writeExtensionConfigs = () => {
         called++
-        return Promise.resolve(config)
+        return Promise.resolve()
       }
 
       appSettings.loadAttachedExtensions = () => { return { testExtension: { path: 'test-extension' } } }
@@ -461,6 +478,7 @@ describe('BackendAction', () => {
 
     it('should pass true to StepExecutor\'s "inspect" constructor argument when called with --inspect', async () => {
       subjectUnderTest.writeExtensionConfigs = () => {}
+      subjectUnderTest.pushHooks = () => {}
       subjectUnderTest._startSubProcess = async function () {
         if (this.backendProcess.executor.inspect !== true) throw new Error('Expected third constructor argument "inspect" to true.')
       }
@@ -469,6 +487,7 @@ describe('BackendAction', () => {
     })
 
     it('should pass false to StepExecutor\'s "inspect" constructor argument when called without --inspect', async () => {
+      subjectUnderTest.pushHooks = () => {}
       subjectUnderTest._startSubProcess = async function () {
         if (this.backendProcess.executor.inspect !== false) throw new Error('Expected third constructor argument "inspect" to false.')
       }