+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v1.5.0
 * Added `extension manage` command to attach/detache multiple extensions at once
-* Updated to use `@shopgate/cloud-sdk-webpack@^1.14.0`
+* Updated to use `@shopgate/cloud-sdk-webpack@^1.14.1`
 
 ## v1.4.1
 * Fix filename of pipeline in boilerplate