+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.5.0
 * Added `extension manage` command to attach/detache multiple extensions at once
 * Updated to use `@shopgate/cloud-sdk-webpack@^1.14.0`
+* Updated `init` command to not ask for permission if option `--force` is used
 
 ## v1.4.1
 * Fix filename of pipeline in boilerplate