+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v1.5.0
 * Added `extension manage` command to attach/detache multiple extensions at once
+* Added ability to log request and response to console when using context.tracedRequest()
 * Updated `init` command to not ask for permission if option `--force` is used
 * Updated to use `@shopgate/cloud-sdk-webpack@^1.14.2`