+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Updated `init` command to not ask for permission if option `--force` is used
 * Updated to use `@shopgate/cloud-sdk-webpack@^1.14.2`
 * Fixed behaviour of --trusted on `extension create` (leaving this flag without trusted|false will create a trusted extension)
+* Added `backend start` will reset hooks
 
 ## v1.4.1
 * Fix filename of pipeline in boilerplate