+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added ability to log request and response to console when using context.tracedRequest()
 * Updated `init` command to not ask for permission if option `--force` is used
 * Updated to use `@shopgate/cloud-sdk-webpack@^1.14.2`
+* Fixed behaviour of --trusted on `extension create` (leaving this flag without trusted|false will create a trusted extension)
 
 ## v1.4.1
 * Fix filename of pipeline in boilerplate