+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * Added `extension manage` command to attach/detache multiple extensions at once
 * Added ability to log request and response to console when using context.tracedRequest()
 * Updated `init` command to not ask for permission if option `--force` is used
-* Updated to use `@shopgate/cloud-sdk-webpack@^1.15.1`
+* Updated to use `@shopgate/cloud-sdk-webpack@^1.15.2`
 * Fixed behaviour of --trusted on `extension create` (leaving this flag without trusted|false will create a trusted extension)
 
 ## v1.4.1