+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Updated `init` command to not ask for permission if option `--force` is used
 * Updated to use `@shopgate/cloud-sdk-webpack@^1.15.2`
 * Fixed behaviour of --trusted on `extension create` (leaving this flag without trusted|false will create a trusted extension)
+* Added support for hooks in pipelines
+* Fixed crash while activating local pipelines on the remote server
 
 ## v1.4.1
 * Fix filename of pipeline in boilerplate