+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed crash while activating local pipelines on the remote server
 * Removed resetting the app's start page URL upon `frontend setup`
 * Add error on storage set if the value is undefined (adapt to production behaviour)
+* Added `backend start` will reset hooks
 
 ## v1.4.1
 * Fix filename of pipeline in boilerplate