+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.6.0
 * Promisified version of `tracedRequest` method
-* Upgraded development dependencies
+* Upgraded vulnerable dependencies
+* Improved error message on invalid extension-config
 
 ## v1.5.0
 * Added `extension manage` command to attach/detache multiple extensions at once