+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v1.6.0
 * Promisified version of `tracedRequest` method
 * Upgraded vulnerable dependencies
+* Improved error message on invalid extension-config
+* Extension-config validated during `backend start`
 
 ## v1.5.0
 * Added `extension manage` command to attach/detache multiple extensions at once