+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Upgraded vulnerable dependencies
 * Improved error message on invalid extension-config
 * Extension-config validated during `backend start`
+* Fix bug setting undefined values in storages
 
 ## v1.5.0
 * Added `extension manage` command to attach/detache multiple extensions at once